refactored methods in EventConcern
