+++ app/models/intake.rb
@@ -2,6 +2,7 @@
 
 class Intake < CaseflowRecord
   class FormTypeNotSupported < StandardError; end
+  include EventConcern
 
   belongs_to :user
   belongs_to :veteran
@@ -268,23 +269,6 @@ class Intake < CaseflowRecord
   # rubocop:enable Metrics/AbcSize
   # rubocop:enable Metrics/MethodLength
 
-  # Check if this Intake is associated with a DecisionReviewCreatedEvent
-  def from_decision_review_created_event?
-    if from_event?
-      # retrieve the record and the event the record is tied to
-      event = event_record.event
-
-      event.type == DecisionReviewCreatedEvent.name
-    else
-      return false
-    end
-  end
-
-  # Check if this Intake is associated with any Event, regardless of type
-  # check if this Intake exists in the Event Records table
-  def from_event?
-    event_record.present?
-  end
 
   # Optionally implement this methods in subclass
   def validate_detail_on_start
