+++ app/models/user.rb
@@ -2,6 +2,7 @@
 
 class User < CaseflowRecord # rubocop:disable Metrics/ClassLength
   include BgsService
+  include EventConcern
 
   has_many :dispatch_tasks, class_name: "Dispatch::Task"
   has_many :document_views
@@ -473,25 +474,6 @@ class User < CaseflowRecord # rubocop:disable Metrics/ClassLength
 
   def show_reader_link_column?
     false
-  end
-
-  # Check if this User is associated with a DecisionReviewCreatedEvent
-  def from_decision_review_created_event?
-    if from_event?
-      # retrieve the record and the event the record is tied to
-      record = EventRecord.find_by(backfill_record_id: id)
-      event = Event.find(record.event_id)
-
-      event.type == "DecisionReviewCreatedEvent"
-    else
-      return false
-    end
-  end
-
-  # Check if this User is associated with any Event, regardless of type
-  # check if this User exists in the Event Records table
-  def from_event?
-    EventRecord.find_by(backfill_record_id: id).present?
   end
 
   private
