+++ app/models/veteran.rb
@@ -6,6 +6,7 @@
 #       VACOLS vet values (coming from Appeal#veteran_full_name, etc)
 class Veteran < CaseflowRecord
   include AssociatedBgsRecord
+  include EventConcern
 
   has_many :available_hearing_locations,
            foreign_key: :veteran_file_number,
@@ -203,24 +204,6 @@ class Veteran < CaseflowRecord
     service.map { |service| service[:pay_grade] }.compact
   end
 
-  # Check if this Veteran is associated with a DecisionReviewCreatedEvent
-  def from_decision_review_created_event?
-    if from_event?
-      # retrieve the record and the event the record is tied to
-      record = EventRecord.find_by(backfill_record_id: id)
-      event = Event.find(record.event_id)
-
-      event.type == "DecisionReviewCreatedEvent"
-    else
-      return false
-    end
-  end
-
-  # Check if this Veteran is associated with any Event, regardless of type
-  # check if this Veteran exists in the Event Records table
-  def from_event?
-    EventRecord.find_by(backfill_record_id: id).present?
-  end
 
   alias zip zip_code
   alias address_line_1 address_line1
